fix: has_object_permission for project

+++ softdesk_management/views.py
@@ -50,9 +50,9 @@ class ProjectAPIView(APIView):
                 },
                 status=400)
 
-    @staticmethod
-    def get(request, project_id):
-        project = get_object_or_404(Project, pk=project_id)
+    def get(self, request, project_id):
+        project = Project.objects.get(pk=project_id)
+        self.check_object_permissions(request, project)
         return JsonResponse(
             {
                 "project_id": project.pk,
@@ -64,10 +64,10 @@ class ProjectAPIView(APIView):
             status=200
         )
 
-    @staticmethod
-    def put(request, project_id):
+    def put(self, request, project_id):
         try:
             project = get_object_or_404(Project, pk=project_id)
+            self.check_object_permissions(request, project)
             data = JSONParser().parse(request)
             serializer = serializers.ProjectSerializer(project, data=data)
             if serializer.is_valid(raise_exception=True):
@@ -91,6 +91,7 @@ class ProjectAPIView(APIView):
     def delete(self, request, project_id):
         try:
             project = get_object_or_404(Project, pk=project_id)
+            self.check_object_permissions(request, project)
             project.delete()
             return JsonResponse(
                 {
